Create Calculation and Transaction

+++ AraratBankRatesAPI/Models/DTO/TransactionResponse.cs
@@ -7,9 +7,9 @@ namespace AraratBankRatesAPI.Models.DTO
         public int Id { get; set; }
         public DateTime CreatedDate { get; set; }
         public Exchange Exchange { get; set; }
-        public decimal ExchangeRate { get; set; }
-        public decimal GivenAmount { get; set; }
-        public decimal ReceivedAmount { get; set; }
+        public double ExchangeRate { get; set; }
+        public double GivenAmount { get; set; }
+        public double ReceivedAmount { get; set; }
         public string TransactionStatus { get; set; }
     }
 }
