+++ AraratBankRatesAPI/Models/Domain/Transaction.cs
@@ -6,9 +6,9 @@
         public DateTime CreatedDate { get; set; }
         public int ExchangeId { get; set; }
         public Exchange Exchange { get; set; }
-        public decimal ExchangeRate { get; set; }
-        public decimal GivenAmount  { get; set; }
-        public decimal ReceivedAmount { get; set; }
+        public double ExchangeRate { get; set; }
+        public double GivenAmount  { get; set; }
+        public double ReceivedAmount { get; set; }
         public string TransactionStatus { get; set; }
         public string UserId { get; set; }
         public ApplicationUser ApplicationUser { get; set; }
